+++ src/app/grid/normal/normal.component.ts
@@ -1,5 +1,5 @@
-import { Component, HostBinding, Input, ViewEncapsulation } from '@angular/core';
-import { Pokemon } from '@shared/interfaces/pokemon';
+import { Component, HostBinding, ViewEncapsulation } from '@angular/core';
+import { GridBaseAppearanceDirective } from '../components/grid-base-appearance.directive';
 
 @Component({
   selector: 'app-normal',
@@ -7,24 +7,8 @@ import { Pokemon } from '@shared/interfaces/pokemon';
   styleUrls: ['./normal.component.scss'],
   encapsulation: ViewEncapsulation.None
 })
-export class NormalComponent {
+export class NormalComponent extends GridBaseAppearanceDirective {
 
   @HostBinding('class') classes = 'grid-item';
-
-  _pokemon: Pokemon | undefined;
-
-  @Input()
-  public get pokemon(): Pokemon | undefined {
-    return this._pokemon;
-  }
-
-  public set pokemon(pokemon: Pokemon | undefined) {
-    this._pokemon = pokemon;
-    // when the pokemon is set check it and set class
-    this.isInactive = !this._pokemon?.isOwned;
-  }
-
-
-  @HostBinding('class.inactive') isInactive: boolean = !this.pokemon?.isOwned;
 
 }