+++ src/app/shared/enums/pokemon-category.enum.ts
@@ -1,0 +1,22 @@
+export enum PokemonCategory {
+  UNCATEGORIZED = 'uncategorized',
+  STARTER = 'starter',
+  PSEUDO_LEGENDARY = 'pseudo-legendary',
+  ULTRA_BEAST = 'ultra-beast',
+  SUB_LEGENDARY = 'sub-legendary',
+  LEGENDARY = 'legendary',
+  MYSTICAL = 'mystical',
+  FOSSIL = 'fossil',
+  BABY = 'baby',
+
+  REGIONAL = 'regional',
+  ALOLA = 'alola',
+  GALAR = 'galar',
+  HISUI = 'hisui',
+
+  UNOWN = 'unown',
+  ALCREMIE = 'alcremie',
+  VIVILLON = 'vivillon',
+
+  GIGANTAMAX = 'gmax',
+}