+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.5.0 - 2022-03-20
+
+### Added
+- Filter for pokemon categories like legendary or fossil
+
+### Changed
+- Changed "hide inactives" filter to owned status filter
+- Fixed Diancie-Mega bankable flag being true
+- Fixed detail appearance of grid showing empty moves
+- Fixed missing icon for RNGs
+- Fixed missing Avalugg entry
+
+
 ## 0.4.2 - 2022-03-19
 ### Changed
 - Updated NPM-Packages